organize command files into one folder
